+++ MainLib/IDictionary.h
@@ -26,7 +26,7 @@ namespace Hlib
         virtual ET_ReturnCode eGetLexemesByInitialForm (const CEString&) = 0;
         virtual ET_ReturnCode eGenerateAllForms () = 0;
         virtual ET_ReturnCode eGenerateFormsForSelectedLexemes() = 0;
-        virtual ET_ReturnCode eCountLexemes (int& iLexemes) = 0;
+        virtual ET_ReturnCode eCountLexemes (int64_t& iLexemes) = 0;
 
         virtual ET_ReturnCode eGetFirstLexeme (ILexeme *& pLexeme) = 0;
         virtual ET_ReturnCode eGetNextLexeme (ILexeme *& pLexeme) = 0;
